WASM: Move WASMDecoder class to separate file
Define and use WATGenerator class in wasm_to_wat.cpp
